training-ai-assistant/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==0.8.11
Markdown==3.5.1
sentence-transformers==2.2.2
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
faiss-cpu>=1.9.0
